Update Server Deletion condition

+++ server/app/controllers/server/index.js
@@ -1,5 +1,6 @@
 const db = require("../../models");
 const Server = db.servers;
+const Task = db.tasks;
 const Op = db.Sequelize.Op;
 const { uniqueNamesGenerator, adjectives, colors, animals } = require('unique-names-generator');
 
@@ -71,23 +72,31 @@ exports.findOne = (req, res) => {
 // TODO delete only if none task is processing by the current server
 exports.delete = (req, res) => {
     const id = req.params.id;
-    Server.destroy({
-        where: { id: id }
-    })
-        .then(num => {
-            if (num === 1) {
-                res.send({
-                    message: "Server was deleted successfully!"
+    Task.count().then(c => {
+        if (c === 0) {
+            Server.destroy({
+                where: { id: id }
+            })
+                .then(num => {
+                    if (num === 1) {
+                        res.send({
+                            message: "Server was deleted successfully!"
+                        });
+                    } else {
+                        res.send({
+                            message: `Cannot delete Server with id=${id}. Maybe Task was not found!`
+                        });
+                    }
+                })
+                .catch(err => {
+                    res.status(500).send({
+                        message: "Could not delete Task with id=" + id
+                    });
                 });
-            } else {
-                res.send({
-                    message: `Cannot delete Server with id=${id}. Maybe Task was not found!`
-                });
-            }
-        })
-        .catch(err => {
-            res.status(500).send({
-                message: "Could not delete Task with id=" + id
+        } else {
+            res.status(400).send({
+                message: `Cannot delete Server with id=${id}. Maybe Server not found!`
             });
-        });
+        }
+    })
 };
